requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
flask>=3.0.0
pydantic>=2.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
werkzeug>=3.0.0
accelerate>=0.24.0
safetensors>=0.4.0

